Added implementation to IPortfolio interface. Made a program which instantiates all the interfaces with their implementation, and adds some stocks to the portfolio and then prints it.

+++ Lektion6/Stocks/Stocks/Interfaces/IPortfolio.cs
@@ -9,7 +9,7 @@ namespace Stocks.Interfaces
         public string StockName { get; set; }
         public string StockSymbol { get; set; }
         public int NumberOfStocks { get; set; }
-        public int StockValue { get; set; }
+        public double StockValue { get; set; }
     }
